fix max values update test
