+++ src/components/main/liquidity/LiquidityTokenItem.style.sass
@@ -24,5 +24,17 @@
         font-size: 15px
         letter-spacing: 0.5px
         color: $textBlack
+        margin-bottom: 4px
+      .sub-title
+        font-style: normal
+        font-weight: 400
+        font-size: 15px
+        letter-spacing: 0.5px
+        color: $grey-text
+        margin-top: 4px
+      .sub-row
+        display: flex
+        flex-direction: column
+        align-items: flex-end
   .avatar
     margin-left: 4px
