+++ src/components/main/liquidity/LiquidityTokenItem.tsx
@@ -4,6 +4,7 @@ import "./LiquidityTokenItem.style.sass"
 import {BorrowSupplyItem} from "models/types"
 import {beautifyNumber} from "utils/utils"
 import {icons} from "utils/icons"
+import Big from "big.js"
 
 export interface LiquidityTokenItemProps {
   item: BorrowSupplyItem
@@ -25,7 +26,10 @@ export const LiquidityTokenItem = ({
         <Avatar className="avatar" size={30} icon={icons[item.symbol]}/>
         <div className="row">
           <span className="title">{item.name}</span>
-          <span className="title">{beautifyNumber(item.liquidity, true)}</span>
+          <div className="sub-row">
+            <span className="title">{beautifyNumber(item.liquidity, true)}</span>
+            <span className="sub-title">{Big(item.liquidity / item.tokenUsdValue).toFixed(2)}</span>
+          </div>
         </div>
       </div>
     </div>
