+++ src/screens/transaction/TransactionViewModel.ts
@@ -18,7 +18,7 @@ import {BUSD} from "models/BUSD"
 import {TRANSACTION_STATUS, transactionStore} from "stores/app/transactionStore"
 import {NavigateFunction} from "react-router-dom"
 import AutosizeInput from "react-input-autosize"
-import {convertValue, cutString, DIGITS_INPUT, NUMBER} from "utils/common"
+import {convertValue, DIGITS_INPUT} from "utils/common"
 
 export class TransactionViewModel {
   item = {} as BorrowSupplyItem
@@ -82,10 +82,13 @@ export class TransactionViewModel {
     }
 
     this.swapReaction = reaction(() => this.inputFiat, (val) => {
-      this.inputValue = !val ?
-        this.inputValueToken ? this.inputValueToken.toString() : "" :
-        this.inputValueFiat ? this.inputValueFiat.toString(): ""
-      this.inputRef?.focus()
+      if (val) {
+        // input fiat
+        this.inputValue = this.inputValueFiat ? this.inputValueFiat.toString(): ""
+      } else {
+        // input token
+        this.inputValue = this.inputValueToken ? this.inputValueToken.toString() : ""
+      }
     })
 
     if (getProviderStore.currentAccount) {
@@ -132,9 +135,14 @@ export class TransactionViewModel {
     return this.item.symbol === getProviderStore.currentNetwork.BUSDSymbol
   }
 
+  get safeInputValue() {
+    if (isEmpty(this.inputValue)) return "0"
+    return this.inputValue
+  }
+
   get isEnoughBalance() {
     if (this.isDeposit) {
-      return this.inputFiat ? this.item.balance.mul(this.item.tokenUsdValue).gte(+this.inputValue) : this.item.balance.gte(+this.inputValue)
+      return this.inputFiat ? Big(this.item.balance).mul(this.item.tokenUsdValue).gte(this.safeInputValue) : Big(this.item.balance).gte(this.safeInputValue)
     }
     return true
   }
@@ -288,7 +296,7 @@ export class TransactionViewModel {
   }
 
   get inputValueTOKEN() {
-    return this.inputFiat ? +this.inputValueToken : +this.inputValue
+    return Big(this.inputFiat ? this.inputValueToken : this.safeInputValue)
   }
 
   get getNewBorrowLimit() {
@@ -301,10 +309,10 @@ export class TransactionViewModel {
       if (!this.item.isEnteredTheMarket) return this.borrowLimit
 
       if (this.isWithdraw) {
-        return this.borrowLimit - this.inputValueUSD * +this.collateralMantissa
+        return this.borrowLimit - this.inputValueUSD * (+this.collateralMantissa)
       }
 
-      return this.borrowLimit + this.inputValueUSD * +this.collateralMantissa
+      return this.borrowLimit + this.inputValueUSD * (+this.collateralMantissa)
     }
 
     return this.borrowBalance ? this.borrowBalance : 0
@@ -398,18 +406,6 @@ export class TransactionViewModel {
   }
 
   // FOR DEPOSIT
-  get hypotheticalBorrowLimitUsedForDeposit() {
-    if (!this.hypotheticalCollateralSupply || !this.totalBorrow) return 0
-    if (this.hypotheticalCollateralSupply < 0) return 100
-    if (!this.item.isEnteredTheMarket) return this.borrowLimitUsed
-
-    const limit =
-      (this.totalBorrow / this.hypotheticalCollateralSupply) * 100
-
-    return limit > 100 ? 100 : parseFloat(limit.toFixed(2))
-  }
-
-  // FOR DEPOSIT
   get hypotheticalCollateralSupply() {
     if (!this.inputValue) return 0
 
@@ -418,6 +414,20 @@ export class TransactionViewModel {
     }
 
     return this.borrowLimit - this.inputValueUSD
+  }
+
+  // FOR DEPOSIT
+  get hypotheticalBorrowLimitUsedForDeposit() {
+    if (!this.hypotheticalCollateralSupply || !this.totalBorrow) return 0
+
+    if (this.hypotheticalCollateralSupply < 0) return 100
+
+    if (!this.item.isEnteredTheMarket) return this.borrowLimitUsed
+
+    const limit =
+      (this.totalBorrow / this.hypotheticalCollateralSupply) * 100
+
+    return limit > 100 ? 100 : parseFloat(limit.toFixed(2))
   }
 
   get hypotheticalBorrowLimitUsed() {
@@ -472,19 +482,23 @@ export class TransactionViewModel {
     return this.isBorrow || this.isRepay ? "borrow" : ""
   }
 
+  get blabla() {
+    return this.inputFiat ? this.inputValueToken : this.inputValue
+  }
+
   get isMaxValueSet() {
-    let input = this.inputValueTOKEN.toString()
+    let input = String(this.blabla)
 
     if (this.isDeposit) {
-      return input === cutString(this.item.balance.toString())
+      return input === this.item.balance.toString()
     }
 
     if (this.isWithdraw) {
-      return input === cutString(this.item.supply.toString())
+      return input === this.item.supply.toString()
     }
 
     if (this.isRepay) {
-      return input === cutString(this.item.borrow.toString())
+      return input === this.item.borrow.toString()
     }
 
     return false
@@ -553,13 +567,6 @@ export class TransactionViewModel {
           transactionStore.transactionMessageStatus.secondStep.status = TRANSACTION_STATUS.ERROR
         }
       } else if (this.isBorrow) {
-        const isMarketExist = await this.isMarketExist()
-
-        if (!isMarketExist) {
-          Logger.info("Market is not available!!") // TODO show toast
-          return
-        }
-
         transactionStore.transactionMessageVisible = true
         transactionStore.transactionMessageStatus.firstStep.message = t("transaction.borrow")
         transactionStore.transactionMessageStatus.secondStep.visible = false
@@ -584,7 +591,7 @@ export class TransactionViewModel {
         if (
           Big(+inputValue)
             .times(1e18)
-            .div(this.item.exchangeRateCurrent)
+            .div(this.item.exchangeRateCurrent) // TODO check
             .lt(1)
         ) {
           Logger.info("error")
@@ -598,6 +605,7 @@ export class TransactionViewModel {
 
         try {
           let valueToSend = this.isMaxValueSet ? this.item.supplyBalance : inputValue
+
           const {hash} = await this.cTokenContract.withdraw(valueToSend)
 
           if (hash) {
@@ -714,6 +722,7 @@ export class TransactionViewModel {
     }
 
     if (this.isWithdraw) {
+      // save value
       maxValue = this.inputFiat ? Big(this.item.tokenUsdValue).mul(this.item.supply) : this.item.supply
     }
 
@@ -721,7 +730,7 @@ export class TransactionViewModel {
       if (this.item.balance.eq(0)) {
         maxValue = 0
       } else {
-        maxValue = +this.item.borrow ? +this.item.borrow : 0
+        maxValue = this.item.borrow ? this.item.borrow : 0
 
         if (this.inputFiat) {
           maxValue = Big(maxValue).mul(this.item.tokenUsdValue)
@@ -729,14 +738,9 @@ export class TransactionViewModel {
       }
     }
 
-    const isInvalidValue = !new RegExp(NUMBER).test(maxValue);
-
     this.setInputValue(
-      isInvalidValue
-        ? cutString(String(maxValue))
-        : String(maxValue) || "0"
+      String(maxValue) || "0"
     )
-    this.inputRef?.focus()
   }
 
   onSwap = () => {
